Fix code duplication
